+++ lib/nretnil-translate.rb
@@ -2,6 +2,7 @@
 # encoding: UTF-8
 # frozen_string_literal: true
 
+require 'nretnil/translate/parseltongue'
 require 'nretnil/translate/pig_latin'
 require 'nretnil/translate/scramble'
 require 'nretnil/translate/version'